Method descriptions + upgrade command
I've done some javadoc description on all the methods I needed to get a small overview around.

Personally I couldn't find a read-thread in the program until i did so. Even though Henrik said it "wasn't needed", I found it to help emendously!

I've also made it possible to use "upgrade" as a command word, and done some minor inclusions of how to process the keyword aswell. Included a new keyword that needs to be accounted for, "back", which is probably gonna be very useful in the end game.